requirements.txt
# Лабораторные работы по компьютерному зрению
# Зависимости проекта
# Используем гибкие версии для совместимости с новыми версиями Python

# Основные библиотеки для обработки изображений
opencv-python>=4.8.0
numpy>=1.24.0

# Визуализация
matplotlib>=3.7.0

# GUI (графический интерфейс)
PyQt5>=5.15.0

# Дополнительные алгоритмы обработки изображений
scikit-image>=0.21.0

# Jupyter Notebook (для отчётов)
jupyter>=1.0.0
notebook>=7.0.0
ipykernel>=6.25.0

# Дополнительные зависимости для Jupyter
ipython>=8.14.0
ipywidgets>=8.1.0

# Утилиты
pillow>=10.0.0

# Deep Learning (optional, for future neural network implementation)
# torch>=2.0.0
# torchvision>=0.15.0

